requirements.txt
Django==4.2.7
aiogram==3.3.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
celery==5.3.4
django-redis==5.4.0
Pillow==10.1.0
